+++ ui/packages/shared/components/src/Select/index.tsx
@@ -24,6 +24,7 @@ export interface SelectElement {
 
 export interface SelectItem {
   key: string;
+  disabled?: boolean;
   element: SelectElement;
 }
 
@@ -42,20 +43,28 @@ const Select = ({
   width,
   className = '',
   loading,
+  primary = false,
+  disabled = false,
 }: {
   items: SelectItem[];
   selectedKey: string | undefined;
-  onSelection: (value: string | undefined) => void;
+  onSelection: (value: string) => void;
   placeholder?: string;
   width?: number;
   className?: string;
   loading?: boolean;
+  primary?: boolean;
+  disabled?: boolean;
 }): JSX.Element => {
   const selection = items.find(v => v.key === selectedKey) ?? {
     key: selectedKey,
     element: {active: <>{selectedKey}</>, expanded: <>{selectedKey}</>},
   };
   const {loader} = useParcaContext();
+  const defaultStyles =
+    'relative bg-gray-50 dark:bg-gray-900 border-t border-r border-b border-l dark:border-gray-600 rounded-md shadow-sm pl-3 pr-10 py-2 text-left cursor-default focus:outline-none focus:ring-1 focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm text-sm';
+  const primaryStyles =
+    'relative text-gray-100 dark-gray-900 justify-center bg-indigo-600 border border-indigo-500 font-medium py-2 px-4 rounded-md shadow-sm pl-3 pr-10 py-2 text-left cursor-default focus:outline-none focus:ring-1 focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm text-sm';
 
   return (
     <>
@@ -66,7 +75,8 @@ const Select = ({
               <Listbox.Button
                 className={cx(
                   width !== undefined ? `w-${width}` : '',
-                  'relative bg-gray-50 dark:bg-gray-900 border-t border-r border-b border-l dark:border-gray-600 rounded-md shadow-sm pl-3 pr-10 py-2 text-left cursor-default focus:outline-none focus:ring-1 focus:ring-indigo-500 focus:border-indigo-500 sm:text-sm text-sm',
+                  disabled ? 'opacity-50 cursor-not-allowed' : '',
+                  primary ? primaryStyles : defaultStyles,
                   {[className]: className.length > 0}
                 )}
               >
@@ -88,49 +98,62 @@ const Select = ({
                 leaveFrom="opacity-100"
                 leaveTo="opacity-0"
               >
-                <Listbox.Options
-                  className={cx(
-                    width !== undefined ? `w-${width}` : '',
-                    'absolute z-10 mt-1 bg-gray-50 dark:bg-gray-900 dark:border-gray-600 shadow-lg rounded-md py-1 text-base ring-1 ring-black dark:ring-white ring-opacity-5 dark:ring-opacity-20 overflow-auto focus:outline-none sm:text-sm'
-                  )}
-                >
+                <Listbox.Options>
                   {loading === true ? (
-                    <div className="w-[270px]">{loader}</div>
+                    <div className="absolute z-10 mt-1 bg-gray-50 dark:bg-gray-900 dark:border-gray-600 shadow-lg rounded-md py-1 text-base ring-1 ring-black dark:ring-white ring-opacity-5 dark:ring-opacity-20 overflow-auto focus:outline-none sm:text-sm">
+                      <div className="w-[270px]">{loader}</div>
+                    </div>
                   ) : (
-                    items.map(option => (
-                      <Listbox.Option
-                        key={option.key}
-                        className={({active}) =>
-                          cx(
-                            active && 'text-white bg-indigo-600',
-                            'cursor-default select-none relative py-2 pl-3 pr-9'
-                          )
-                        }
-                        value={option.key}
-                      >
-                        {({selected, active}) => (
-                          <>
-                            <div className="flex items-center">
-                              <span
-                                className={cx(selected ? 'font-semibold' : 'font-normal', 'ml-3')}
-                              >
-                                {option.element.expanded}
-                              </span>
-                            </div>
-                            {selected ? (
-                              <span
-                                className={cx(
-                                  active ? 'text-white' : 'text-indigo-600',
-                                  'absolute inset-y-0 right-0 flex items-center pr-4'
-                                )}
-                              >
-                                <CheckIcon className="h-5 w-5" aria-hidden="true" />
-                              </span>
-                            ) : null}
-                          </>
-                        )}
-                      </Listbox.Option>
-                    ))
+                    <>
+                      {items.length > 0 && (
+                        <div
+                          className={cx(
+                            width !== undefined ? `w-${width}` : '',
+                            'absolute z-10 mt-1 bg-gray-50 dark:bg-gray-900 dark:border-gray-600 shadow-lg rounded-md py-1 text-base ring-1 ring-black dark:ring-white ring-opacity-5 dark:ring-opacity-20 overflow-auto focus:outline-none sm:text-sm'
+                          )}
+                        >
+                          {items.map(option => (
+                            <Listbox.Option
+                              key={option.key}
+                              disabled={option.disabled ?? false}
+                              className={({active, disabled}) =>
+                                cx(
+                                  active && 'text-white bg-indigo-600',
+                                  'cursor-default select-none relative py-2 pl-3 pr-9',
+                                  disabled && 'opacity-50'
+                                )
+                              }
+                              value={option.key}
+                            >
+                              {({selected, active}) => (
+                                <>
+                                  <div className="flex items-center">
+                                    <span
+                                      className={cx(
+                                        selected ? 'font-semibold' : 'font-normal',
+                                        'ml-3'
+                                      )}
+                                    >
+                                      {option.element.expanded}
+                                    </span>
+                                  </div>
+                                  {selected ? (
+                                    <span
+                                      className={cx(
+                                        active ? 'text-white' : 'text-indigo-600',
+                                        'absolute inset-y-0 right-0 flex items-center pr-4'
+                                      )}
+                                    >
+                                      <CheckIcon className="h-5 w-5" aria-hidden="true" />
+                                    </span>
+                                  ) : null}
+                                </>
+                              )}
+                            </Listbox.Option>
+                          ))}
+                        </div>
+                      )}
+                    </>
                   )}
                 </Listbox.Options>
               </Transition>