+++ ui/packages/shared/functions/src/index.ts
@@ -15,6 +15,14 @@ import format from 'date-fns/format';
 import {Label} from '@parca/client';
 import colors from 'tailwindcss/colors';
 
+export * from './hooks';
+
+export type NavigateFunction = (
+  path: string,
+  queryParams: any,
+  options?: {replace?: boolean}
+) => void;
+
 export const SEARCH_STRING_COLOR = '#e39c9c';
 
 export const capitalize = (a: string): string =>
@@ -192,7 +200,7 @@ export const getLastItem = (thePath: string | undefined): string | undefined => 
 
 const transformToArray = (params: string): string[] => params.split(',');
 
-export const parseParams = (querystring: string): Record<string, string | string[]> => {
+export const parseParams = (querystring: string): Record<string, string | string[] | undefined> => {
   const params = new URLSearchParams(querystring);
 
   const obj: Record<string, string | string[]> = {};
@@ -212,9 +220,27 @@ export const parseParams = (querystring: string): Record<string, string | string
   return obj;
 };
 
-export const convertToQueryParams = (params: Record<string, string>): string =>
+export const selectQueryParam = (key: string): string | string[] | undefined => {
+  const router = parseParams(window.location.search);
+
+  if (key === 'dashboard_items') {
+    let dashboardItems = router[key];
+    if (typeof dashboardItems === 'string') {
+      dashboardItems = [dashboardItems] ?? [];
+    }
+    return dashboardItems;
+  }
+
+  if (key === 'compare_a' || key === 'compare_b') {
+    return router[key] === 'true' ? 'true' : 'false';
+  }
+
+  return router[key];
+};
+
+export const convertToQueryParams = (params: {[key: string]: string | string[]}): string =>
   Object.keys(params)
-    .map(key => key + '=' + params[key])
+    .map((key: string) => `${key}=${params[key] as string}`)
     .join('&');
 
 export function convertUTCToLocalDate(date: Date): Date {