+++ ui/packages/shared/icons/src/CloseIcon.tsx
@@ -19,11 +19,11 @@ const CloseIcon = (): JSX.Element => {
         width="24"
         height="24"
         viewBox="0 0 24 24"
-        stroke-width="2"
+        strokeWidth="2"
         stroke="currentColor"
         fill="none"
-        stroke-linecap="round"
-        stroke-linejoin="round"
+        strokeLinecap="round"
+        strokeLinejoin="round"
       >
         <path stroke="none" d="M0 0h24v24H0z" />
         <line x1="18" y1="6" x2="6" y2="18" />