+++ ui/packages/shared/functions/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.16.56 (2023-01-11)
+
+**Note:** Version bump only for package @parca/functions
+
 ## 0.16.55 (2022-12-22)
 
 **Note:** Version bump only for package @parca/functions